move bindings to it's own module create index.ts file so we don't need to import via bindings/bindings
