+++ src/bindings/index.ts
@@ -1,0 +1,1 @@
+export {getBindings} from './bindings';
